py/pillowLab/getGPS/requirements.txt
Pillow
BeautifulSoup4
